backend/requirements.txt
#application
fastapi>=0.61.2
uvicorn>=0.11.3
pydantic>=1.4
email-validator>=1.1.1
python-multipart>=0.0.5
websockets>=8.1
sse-starlette>=0.6.1

#database
databases[postgesql]>=0.3.1
asyncpg>=0.21.0
SQLAlchemy>=1.3.16
alembic>=1.4.2
psycopg2>=2.8.6
pytz>=2020.4

#file handling
aiofiles>=0.6.0

#development
pytest>=5.4.2
pytest-asyncio>=0.12.0
pytest-xdist>=1.32.0
httpx>=0.12.1
asgi-lifespan>=1.0.0
docker>=4.2.0

#authentication
pyjwt>=1.7.1
passlib[bcrypt]>=1.7.2